+++ src/main/java/roomescape/reservation/application/ReservationService.java
@@ -15,8 +15,8 @@ import roomescape.reservation.application.dto.request.ReservationSaveRequest;
 import roomescape.reservation.application.dto.response.ReservationDetailFindResponse;
 import roomescape.reservation.application.dto.response.ReservationSaveResponse;
 import roomescape.reservation.application.port.out.ReservationRepository;
-import roomescape.slot.SlotOccupancy;
-import roomescape.slot.Slot;
+import roomescape.slot.domain.SlotOccupancy;
+import roomescape.slot.domain.Slot;
 import roomescape.slot.application.SlotAssembler;
 import roomescape.waiting.domain.Waiting;
 import roomescape.waiting.domain.WaitingLine;
