+++ src/main/java/roomescape/slot/application/dto/response/SlotFindResponse.java
@@ -1,6 +1,6 @@
-package roomescape.slot.dto.response;
+package roomescape.slot.application.dto.response;
 
-import roomescape.slot.Slot;
+import roomescape.slot.domain.Slot;
 
 import java.time.LocalDate;
 import java.util.List;
