+++ src/main/java/roomescape/waiting/application/WaitingService.java
@@ -6,8 +6,8 @@ import org.springframework.transaction.annotation.Transactional;
 import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
 import roomescape.reservation.application.port.out.ReservationRepository;
-import roomescape.slot.Slot;
-import roomescape.slot.SlotOccupancy;
+import roomescape.slot.domain.Slot;
+import roomescape.slot.domain.SlotOccupancy;
 import roomescape.slot.application.SlotAssembler;
 import roomescape.waiting.domain.Waiting;
 import roomescape.waiting.domain.WaitingLine;
