+++ src/test/java/roomescape/waiting/WaitingServiceTest.java
@@ -10,7 +10,7 @@ import roomescape.exception.ErrorCode;
 import roomescape.exception.EscapeRoomException;
 import roomescape.reservation.application.port.out.ReservationRepository;
 import roomescape.reservationtime.ReservationTime;
-import roomescape.slot.Slot;
+import roomescape.slot.domain.Slot;
 import roomescape.slot.application.SlotAssembler;
 import roomescape.theme.Theme;
 import roomescape.waiting.application.WaitingService;
